Refactor template validation into domain modules
- separate Agent Skills, Agent Plugins, and client-format validation
- replace global validation state with an explicit per-run context
- keep the CLI focused on reporting results and exit status
- split regression coverage into focused domain and integration tests
