+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "test": "node --test scripts/*.test.mjs",
+    "test": "node --test scripts/*.test.mjs scripts/validation/*.test.mjs",
     "validate": "node scripts/validate-template.mjs"
   },
   "devDependencies": {
